// frontend/aircrat-management1/src/app/core/models/aircraft.model.ts
export class Aircraft {
  id?: number;
  aircraftCode?: string;
  aircraftName?: string;
  modelName?: string;
  registrationNumber?: string;
  capacity?: number;
  status?: string;
  manufacturer?: string;
  aircraftType?: string;
  cabinClasses?: string;
  rangeKm?: number;
  cruiseSpeedKmh?: number;
  imageUrl?: string;
}